microservices/sample_service/requirements.txt
tornado==6.0.3
google-cloud-firestore>=1.6.2
google-api-python-client==1.7.11
google-auth>=1.6.2
firebase-admin==3.2.0
jsonschema==3.2.0
redis==3.4.1
google-cloud-core>=1.3.0
